+++ fe/src/lib/components/explorer/Filters.svelte
@@ -1,61 +1,12 @@
 <script>
-	import {isObjNotEmpty} from  '@svizzle/utils'
 	import * as _ from 'lamb';
-	import {RISON} from 'rison2';
 
 	import CategorySelector
 		from '$lib/components/explorer/CategorySelector.svelte';
 	import RangeSlider from '$lib/components/svizzle/RangeSlider.svelte';
 	import Scroller from '$lib/components/svizzle/Scroller.svelte';
-	import {explorerActor} from '$lib/statechart/index.js';
 	import {_staticData} from '$lib/stores/data.js';
 	import {_filters} from '$lib/stores/filters.js';
-	import {_rangeSlidersTheme} from '$lib/stores/theme.js';
-	import {pluckKey} from '$lib/utils/svizzle/utils.js';
-
-	const getSelectedCats = _.pipe([
-		_.filterWith(_.hasKeyValue('selected', true)),
-		pluckKey
-	]);
-
-	const getFiltertQuery = filters => {
-		if (!filters) return null;
-		const query = {};
-		filters.forEach(({values: metrics}) => {
-			_.forEach(
-				metrics,
-				metric => {
-					if (metric.type === 'number' || metric.type === 'date') {
-						const subQuery = {};
-						if (metric.max !== metric.Max) {
-							subQuery['lte'] = metric.max;
-						}
-						if (metric.min !== metric.Min) {
-							subQuery['gte'] = metric.min;
-						}
-						if (isObjNotEmpty(subQuery)) {
-							query[metric.id] = subQuery;
-						}
-					} else if (metric.type === 'category') {
-						if (_.someIn(metric.values, ({selected}) => !selected)) {
-							query[metric.id] = getSelectedCats(metric.values);
-						}
-					}
-				}
-			);
-		});
-		return isObjNotEmpty(query) ? RISON.stringify(query) : '';
-	}
-
-	let lastFilterQuery;
-	$: filterQuery = getFiltertQuery($_filters);
-	$: if (filterQuery !== lastFilterQuery) {
-		explorerActor.send({
-			type: 'SELECTION_CHANGED',
-			newValues: {filterQuery}
-		});
-		lastFilterQuery = filterQuery;
-	}
 </script>
 
 {#if $_filters}
